Finnaly Fixed Connect and Disconnect
The Program however is still a bunch of spaghetti code and althought its the choosen structure.
Honestly this is just a way for me to learn posix complaint shell scripting although if you are seeing this commit its most likely because i've made it "public". Anyway I might revisit this project to give it some optimizations.
i believe that for now the program is more robust although less polished than the previous one (since the original used Bluetoothctl, while my fork uses busctl)

Also thanks to the SO user which gave me the tip to use busctl which although initially it was pretty vague the manual and the program is well made so was easy to see how everything works and to understand a more low level of it 

I hope you don't cringe at my 30'000 errors on the description above and for now i have to eat, it wasn't easy writting this when i have like 30 flies just annoying me